course/pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch>=9.1.1",
    "faiss-cpu>=1.12.0",
    "huggingface-hub>=0.35.3",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.78",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.35",
    "langchain-text-splitters>=0.3.11",
    "minsearch>=0.0.7",
    "openai>=2.2.0",
    "openai-agents>=0.4.0",
    "pydantic-ai>=1.6.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "toyaikit>=0.0.4",
    "tqdm>=4.67.1",
    "youtube-transcript-api>=1.2.2",
]
